[CHERRY-PICK] Adding '34ca6cf382e6f16ca7218dcb2db5e2ca422f305a' to 'CHERRY_PICK_main_2024-02-22_16.26.31' using the python script version 'v1.0.6'.

+++ PresentationLayer/ACCOUNT/AccountingVouchers.aspx.cs
@@ -17288,8 +17288,8 @@ public partial class AccountingVouchers : System.Web.UI.Page
             {
                 if(Convert.ToInt32(txtAcc.Text.ToString().Trim().Split('*')[1])==Convert.ToInt32(PartyNo.Value))
                 {
-                    totalCCAmount = totalCCAmount + Convert.ToInt32(lblAmount.Text) + Convert.ToDouble(txtccAmount.Text);
-                    if (Convert.ToInt32(totalCCAmount) > Convert.ToDouble(txtTranAmt.Text))
+                    totalCCAmount = totalCCAmount + Convert.ToDouble(lblAmount.Text) + Convert.ToDouble(txtccAmount.Text);
+                    if (Convert.ToDouble(totalCCAmount) > Convert.ToDouble(txtTranAmt.Text))
                     {
                         ScriptManager.RegisterStartupScript(this, GetType(), "showalert", "alert('The Total Cost Center Amount Should be Less than or Equal to Transaction Amount.');", true);
                         return;
